refactor: rename refresh_notebook_situation to update_notebook
