+++ hasura/metadata/actions.yaml
@@ -36,10 +36,10 @@ actions:
       - role: orientation_manager
       - role: professional
     comment: Récupération du diagnostic individu pole-emploi
-  - name: refresh_notebook_situations_from_pole_emploi
+  - name: update_notebook_from_pole_emploi
     definition:
       kind: synchronous
-      handler: '{{BACKEND_API_ACTION_URL}}/v1/notebooks/refresh-situations-from-pole-emploi'
+      handler: '{{BACKEND_API_ACTION_URL}}/v1/notebooks/update-notebook-from-pole-emploi'
       headers:
         - name: secret-token
           value_from_env: ACTION_SECRET
@@ -152,7 +152,7 @@ custom_types:
     - name: NPSRatingOutput
     - name: UpdateSocioProOutput
     - name: CreateNotebookOutput
-    - name: RefreshNotebookSituationsFromPoleEmploiOutput
+    - name: UpdateNotebookFromPoleEmploiOutput
     - name: PoleEmploiDossierIndividu
     - name: PoleEmploiDossierIndividuBesoinsParDiagnosticIndividu
     - name: PoleEmploiDossierIndividuBesoins
